Pixel Shifting 2 - The Sequel (The Layering) (#749)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds the ability to layershift. By default this is bound to Ctrl + Shift
+ Pageup/Pagedown (old players will need to reset to defaults or
manually set this). Pixelshifting / layershifting while lying down will
change the layer of the player from lying default (3.8) to regular layer
- 0.01 (3.99), allowing for the player to layer above a standing player,
but by default keeping them underneath regardless. Layer shifting is
capped at 0.05 either up or down, reduced by 0.01 for lying down mobs.

That is not all! If a mob is pixel shifted more than ten pixels in any
direction (or 8 or more downwards, to account for the game's
perspective) other mobs will now be able to pass through them in the
same manner as a wall-leaning player. Both this and layering are reset
if you move, and if a mob pixel shifts to no longer be offset by ten or
more pixels they lose the passthroughability.

## Why It's Good For The Game

Layer shifting was requested by Useroth to be a part of the original PR,
but due to some difficulties I was facing code wise, I was not able to
include it at the time. This now adds it, and ties it to a keybind so
that it adds no additional clutter into the UI and can be disabled for
players who do not want the ability to do this. Layershifting allows for
players to determine the order their sprites show up in, allowing for
better visual representations of particular roleplay encounters, such as
two individuals wholesomely cuddling platonitcally.

The ability to pass by pixel-shifted mobs was left out initially due to
concerns I had that it might be gameable, however since then it has been
a feature requested by several people and has become apparent that the
ability to pass by pixel shifted characters would make situations
in-game easier in which somebody wants to get past someone else or share
a tile with them _**for some reason,**_, making the experience less
tedious and confusing for those involved. This allows for a more
immersive way of achieving this than one person lying down for the other
to get onto their tile, and due to the time it takes to pixel shift will
have minimal impact on combat situations.

(Note: while testing layer shifting I found my player sprite to still
render above other objects while standing up regardless of layer, but
found that layers worked correctly for other mobs without ckeys, or
while lying down. This is something that requires live testing to see if
it works correctly with other players. If it does, it works correctly,
if it turns out this issue occurs, then nothing will have changed
regarding layers for players in-game and the other features will still
work while I figure out a fix.)