Changes Templar helmet FOV, removes them from heavy category (#744)
This PR changes the FOV restrictions of the Templar patron-specific
helmets to only block the behind, the same as the visored sallet does.
To balance it out, they're also removed from the 'heavy' category, and
thus lose the added protection from crits that such helmets would offer.

The reason for this change is that these helmets are not adjustable at
all, unlike the Knight helmet- and when your job is to guard a place,
that becomes pretty difficult to do when you can hardly see a thing
around you- and it kind of sucks to take off these awesome looking
helmets 90% of the time. A lot of people that I know play Templars have
expressed the same sentiment of wishing that the view in them wasn't so
restrictive.

This has been tested on my private repo with no errors.