Ports Sleeping bags and Sleeping cloths (#778)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! --> Stonekeeps PR. original authors words "Sleepcloths!
You can now sew bedsheets and sleepcloths. A sleepcloth is basically a
sleeping bag which you can roll up to carry around with you. For the
added convenience it's also very uncomfortable to sleep on, causing you
to get a mood debuff. An even bigger one if you're of noble blood."
https://github.com/Darkrp-community/OpenKeep/pull/96

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. --> "Convenience!
And a better camping experience."