Makes jester str, luc and int rolling a curve (#799)
currently, jester stat rolling is all sorts of messed up. you have...
way too high a chance to roll a 20 or a one. a 20 in strength, no one
can beat you. a 1, you can't actually play the game at all. 20 in int,
you can learn anything. 1 in int - you can't actually craft a single
damn thing, which is annoying if you're trying to make funny shit for
japes.

this changes it to follow this chart **(MINUS FOUR)**. subtract 4 from
this chart (0 counts as a 1) and that will be the probability


![4-chart](https://github.com/user-attachments/assets/6bab4aef-6da7-4173-b291-f83f45eb7c1e)