Ports the Signal Horn from Stonekeep (#781)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->Ports a cool signal horn to rally people. Given to
the Bog Master, the Guard Captain and the Veteran. Uncraftable.
https://github.com/Darkrp-community/OpenKeep/pull/470

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->Fun and great
for getting the attention of guards and boggards alike. or just as a
funny horn the jester steals. Tested locally and it works