Rehudd Rebels (#324)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

This reverts the dehudding of rebels. 

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Not being able to see fellow rebels is insanely lame. 

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

I don't think this is a very far fetched change. This change benifits
new players mostly who aren't used to examining their fellow rebels.
Overall, It'll lead to less confusion and potential RDM tickets.