added an abandoned mining shaft far south-west in the bog + added missing indoor areas (#754)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->Currently the entire south west of the bog is
barren, so i put a majority medium ore, volf infested mining shaft for
the bravest miners to venture into and mine. Following this i will also
be making a PR that sections off a large portion of the mine in the town
to encourage mining outside of the walls of the town, leading to miners
hiring mercs/adventurers to help with transporting goods back to the
town and for safer mining etc, etc. WARNING this is my first ever time
mapping for ss13 so it is probably not that good, do not ridicule me too
much pls. oh yeah i also added some missing indoors zones for the bog
smithy, and some other roofed ruins

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->More reason to
be in the bog, is good


![image](https://github.com/user-attachments/assets/84071b31-7513-4cbf-ab40-13b374539094)

![image](https://github.com/user-attachments/assets/8edc7ea7-f8d4-49a1-8480-2f8ffb35457c)

![image](https://github.com/user-attachments/assets/7046b02b-2476-4fdb-bda8-98d82e8ff876)

Co-authored-by: Useroth <37159550+Useroth@users.noreply.github.com>