Scomstone Added to Sacrifice list for bandit idol + Sellprice increase (#825)
## About The Pull Request

Scomstone (the one heartfelt lord/hand spawn with. It is an emerald ring
that is a scom radio) is added to the idol sacrifice list.
The sellprice is increased from 35 to 250. 50 more than the choker due
to it taking a slot that doesn't stop you from wearing armor, so
naturally it's better.

## Why It's Good For The Game

Shrimple as.