Removes the last scraps of race restriction
I pointed out verminvolk cant be physicians. Ended up noticing theres a LOT of odd jobs they cant be. This PR removes every example of  race restriction from every single job, Idk if theres a tangible reason for any of these to be restricted. If there is commit away. a