Makes the padded dress dyable. (#838)
## About The Pull Request

The padded dress can now be dyed. It has a default color of dark green,
similar to how it currently looks.
The alternate (purple) padded dress' spawn chance is removed from Queen
consort and Lady Heartfelt spawn.

Note: I'm not completely confident in the greyscale, but most colors
seemed to work just fine.

## Why It's Good For The Game
Some people wished for the padded dress to work with dye bins, and it
makes enough sense. If it's working sufficiently, it removes the need
for the alternate padded dress, so this PR goes ahead and removes them
from spawning normally as well.