Gives Priest graveyard key and gives mortician church key (#849)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives the Mortician a Keyring that has the Church Key and the Graveyard
Key on it.

Also gives the Priest's keyring a graveyard key
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
As it stands, the Morticians can't even access their own rooms because
it requires the Church Key to open, and they only spawn with the
graveyard key. This also helps the Priest, so they can unlock the
Graveyard if there are no Morticians.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->