Gives Grenzels a chance to spawn with a Zweihander (#848)
## About The Pull Request

Gives Grenzelhoft mercenaries a chance (1/3) to spawn with a zweihander
and expert swords instead of expert polearms. Also gives them a satchel
because they don't get one and it sucks.

## Why It's Good For The Game

Its cool. Lets you pretend you're a doppelsoldner or landsknecht. On
Stonekeep grenzelhoft mercenaries only get to spawn with a zweihander
which is interesting.