Changes estoc penfactor to 60 (#855)
This brings it up to just above the penetration factor as a spear
thrust. The reasons for this are as follows:

At 40 penfactor, the estoc does not actually penetrate armour
whatsoever. This is kind of lame for a weapon whose whole point is that
it's meant to be able to penetrate armour. While this may be a bit
better penetration than the spear, the spear is an iron weapon, whereas
the estoc weapon requires 2 steel bars, and should thus reflect the
material investment. This is still not anywhere near the insane
penetration levels of a flail- you are not going to be critting anyone
through their armour with this, but you are at least able to soften them
up before you do, which is quite needed on a weapon whose whole point is
to be used against armour.

In addition, a spear allows you to have this penetration factor whilst
also holding a shield, and also having increased range against your
enemy. WIth the estoc, you have a higher material cost, need to twohand
the weapon, and have to be face to face with your enemy.