Makes MAA less ugly (#880)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes the spawn-helmet of the Man at Arms class from a Bascinet to the
Kettle Helmet.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
BECAUSE THE BASCINET IS SO FUCKING UGLY!! NOBODY TAKES MAA SERIOUSLY
BECAUSE THEY HAVE THIS STUPID HELMET THAT MAKES THEIR HEAD LOOK LIKE THE
TIP OF A PENIS!!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->