Misc sprite bundle. (#876)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
A small batch of seven new sprites to hold over until I can get the
other half of the crops done.
Included are resprites of the following: Mincemeat, eggtoast,
peppersteak, spicedeggs, westernspice, easternspice, and a palette shift
for cloth to be more in-line with the bundled appearance.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
For all the food items, they use placeholder /tg/ sprites. I referenced
and/or copied parts of existing items that are used in the crafting of
these to make the transition not quite so abrupt. It's a transformation
of an object, not an entirely new one- at least as I want to see it.
Westernspice is modeled after rosemary and easternspice is modeled after
ginger. I couldn't really tell what either of them were supposed to be,
so these are my best guesses that fit within the universe.
Cloth is very bright white generally. This might adjust the greyscale to
be a bit darker than whatever colour it's supposed to be, but I think it
looks better in the long run.

![image](https://github.com/user-attachments/assets/189cc689-7d02-4fc0-89c7-d49567e8b0b0)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>