Rockhill Map Cleanup (#881)
## About The Pull Request

removed a duplicate closet in the keep dungeon
removed a stray gate in the moat southeast of the keep
removed a hermes mail machine just ... sitting there in the woods in the
bog village
removeda duplicate hermes in the merchant's. The remaining one is by the
goldface
removed 2 duplicate chests in the 'summer house?' west of the town
removed a floating bed east of the keep walls
removed a bunch of floating stuff above the graveyard north of town
chapel
removed some tables and items stuck in cave walls east of the town spawn
boat

## Why It's Good For The Game

Shrimple as.