NEW swashbuckler rogue archetype (#831)
it's like a rogue but more swashbuckley. you get a rapier and shit. felt
we needed a class that was inbetween rogue and warrior and this seems to
fit well. would have named it musketeer but why the fuck would i give it
a musket thats wildly op

please view the code for stats. balance feedback appreciated