Give bandits t2 love potion reward (#685)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives bandits the love potion once they sacrifice 200 mammons worth of
loot to mattheos
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Right now bandits have no way of actually managing prisoners properly or
securing loyalty and due to the larger pop count, the prison isn't
enough for everyone to get a prisoner. With how far into the code its
put, this shouldnt really have an effect until late into the round
though so I am not opposed to putting it earlier.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: GuppyLaxx <25938722+GuppyLaxx@users.noreply.github.com>