gives queen goodlover trait (#904)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
king gets triumphs from taking the time to fuck his wife, or randoms get
rewarded for making the king a cuck, good times either way
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
someone asked for it on discord
https://discord.com/channels/1240735983276654733/1258555111005880520/1280237574727798834
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->