Catch Button fix. (#906)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Very minor recolour of the catch button.
Before:

![imagecatch](https://github.com/user-attachments/assets/1c1d64ef-be9a-47ea-8db5-2cf64eef1aa0)
After:

![imagecatchfixed](https://github.com/user-attachments/assets/bb96c394-fa7c-4817-9adf-7f97d1cffb13)
Have you noticed this yet? No? Now you have! It's going to bother you
until the end of time- so here I am!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Am I the only one who cannot stop seeing this? It's the only button in
the HUD that appears this way, for no discernible reason. This brings it
in line with all the others. It looks like almost someone recoloured it
in half-way then decided not to for Catch in particular. Who knows why-
but with this PR, the HUD is uniform.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>