Gives lady of heartfelt goodlover trait as well (#909)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
I honestly didnt expect the queen to get it but uhhh, yeah this sorta
has the same reasoning.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I dont know man, I guess people cucking the lord of heartfelt would be
pretty funny
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->