Bugfix for #913 (#931)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
#913 created an issue where instead of spawning one full bundle of 10
sticks, it spawned 10 bundles of 2 sticks in each inventory slot,
causing a massive spillover and needed some minor, but still tedious
cleanup.

This fixes that so that morticians now only spawn with one bundle of 2
sticks. Kind of a nerf, but it's still a little something.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Bugfix.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->