Buffs noble pilgrims, removes sexism (#916)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Removes sexism from noble code, now both genders get the same skills and
stat modifiers, buffed their combat skills to 2(high civilian level)
with the exception of polearms IF OLD, which is 3. This is because old
man with a staff is cool as fuck. Also they shouldn't be able to hold
any of the heavier polearms with old debuffs and -1 str stacked on top
of that unless they game it like horc or something. And lastly made it
all races because they are a foreign noble. Slopfix for until someone
makes a drow noble class or a horc noble class.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Noble was kinda a dogshit class, still should be after these changes
with no useful skills and no unique stats (just checked and they get
better bows than warriors 50% of the time... oops. Still shouldn't be an
issue because warrior is melee coded anyways) other than being a lootbag
and having noble blood. Still vastly inferior to other pilgrim classes
in terms of combat so shouldn't be an issue.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->