Allows repairing damaged and broken doors (#919)
![image](https://github.com/user-attachments/assets/e15e6b39-af26-408a-81cc-37331060735e)

- If a door is damaged, it can be healed for half its health by using
ingredient one. Repairing a damaged (not broken) door restores half
health, so a heavily damaged door might need to be repaired twice.
- If a door is broken, it can be repaired by first using ingredient one,
then ingredient two.

- Locks are preserved this way. (Be smart, don't be on the wrong side of
a locked door when you repair it)
- You need at least 1 level in the corresponding skill to start repairs.
- It takes 30 seconds divided by skill level for repairing each stage.
Means a novice carpenter will stand there for a minute in total to
repair a door. This was done to intentionally prevent combat repairs,
and to make higher skill levels more relevant. A normal carpenter will
barely take 15 seconds or less to fix it.

The costs to repair:
- Wooden door / deadbolt door / tavern door: 2 small logs, carpentry
- Reinforced door (With the sight slit): 1 small log and 1 iron ingot,
carpentry
- Stone door: 2 stone, masonry
- Iron Door (Graveyard gate): 2 iron, blacksmithing

(If I missed a door, please do let me know)