Renames Lieutenant to Guard Captain (#932)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Renames the Guard Lieutenant to Guard Captain, and returns the Captain
honorific.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Because why the FUCK was it changed in the first place? It doesn't mesh
with the time period and NOBODY likes it. Guard Captain also actually
gives POWER to the role to actually... LEAD THE TROOPS.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Gup <guppylaxx@gmail.com>