New migrant wave: The Zybantine Slavers (tested and functional) (#789)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->Arriving to the Isles of Enigma, The Zybantine
Slavers from western deserts of Zybantine (same place as the Desert
Riders). The first light-villain(?) migrant wave, including a Zybantine
Slave Master, 4 Zybantynian Mercs (2 with swords, 2 with wips), and an
imported Zybantynian slave!

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->More variety,
and also for fun!

Also this is what I based the Zybantine name and lore on

![image](https://github.com/user-attachments/assets/b97f2b46-0ec2-4de0-9ea8-a6a7e686b929)


![image](https://github.com/user-attachments/assets/961533e8-bb7b-4af5-9963-d4fa0ea55734)


https://github.com/user-attachments/assets/b3193870-569e-4ad5-a3f2-8ff912044112