Apothecary towner (#648)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds the Apothecary as a towner subclass, they start with expert alchemy
and starting tools.
(Open to changes if the role feels underwhelming)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Has been requested and will allow players who want to try alchemy not
need to go around town looting what they need to start.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->