Weaponsmiths have a job now (Adventurers have iron weapons) (#940)
## About The Pull Request

This PR changes almost every instance of steel weapons to iron weapons,
in an adventurers kit. In instances where there simply was not an iron
weapon with the same niche, (cough cough, bastard swords), it was given
the closest I could give it. This PR is aiming for accuracy on the
downgrades, so I didn't play around with touching up on the loadouts
with some of our new toys such as the shortswords you never see.

This PR also changes the steel equipment skeletons can spawn with to
iron (Sword, chain gloves, chain coif, thank you radarseas for pointing
out the armor). Skeletons are still very valuable lootboxes, especially
for smiths, with this change

NOTE: Dwarfwarrior, sentinel, and witchhunter were not changed. All of
these classes have limited slots, and the latter two have rather unique
silver gear so I thought it best. Amazon was also not changed, as there
is currently a PR that does it for them already

Actual changes are as follows:

- Paladin(Paladin subtype): Bastard sword/Steel sword chance to iron
sword.
- Vaquero: Steel dagger to iron dagger
- Bard(Both subtypes): Steel dagger to iron dagger
- Ranger(Gloom Stalker subtype): Steel dagger (special) to irondagger
- Rogue: Steel daggers to iron daggers for Thief, and rapier to iron
sword for Swashbuckler
- Warrior: Warrior subtype now has a chance for a iron messer instead of
a sabre. Monster hunter now gets Iron sword/Spear/Iron woodaxe instead
of Bastardsword/Billhook/Battleaxe

## Why It's Good For The Game

The stark contrast of work between a weaponsmith and an armorsmith is
appalling. This should hopefully even it out a little. Adventurer haters
may now cheer that adventurers are a bit more impoverished for their
roundstart kit. (Sorry adventurer mains, but there is no iron analogue
to the bastard sword or rapier, you'll have to forgive me)