Gives hostages collars instead of the cursed mask (#672)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
See title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The cursed mask sucks, It hides the face and is an old reference to
something I could care less about. Also back when it was implemented it
was implied the prisoner had the mask to hide their nobility but
everyone already knows hostages now are nobles.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->