Metal uniformity. (#955)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changes a LOT of iron/steel sprites to have uniform colour schemes as to
their metals. Contails a range of edits, from minor to heavy, to enforce
this uniformity. Also changes the colour of steel ingots to match the
weapons and tools made from them.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Considering the material a weapon is made of is a determining factor for
its quality, this serves as a visual indicator of that. Before the
Stonekeep weapon PR, *some* of the older Roguetown sprites followed this
philosophy. This is my attempt to return to that, using the better
shapes of Stonekeep sprites where they're warranted.
![Screenshot 2024-09-09
201507](https://github.com/user-attachments/assets/593bd4bf-c3ff-4fef-aca1-d041a91c3bce)
![Screenshot 2024-09-09
201633](https://github.com/user-attachments/assets/5278bd38-8660-42c8-a685-53722a5e03e2)
![Screenshot 2024-09-09
201650](https://github.com/user-attachments/assets/9ecdbedf-5954-458f-92a4-be7de84f39c5)
I may attempt to change the tinting of iron to red in the future, but
this contains the green tinted sprites. Uniformity first, palette shift
later.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>