Add support for the Auxtools Debugger (#936)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This PR adds support for the auxtools debugger. All that a user needs to
do is have the DM Langserver client extension installed(which comes with
the recommended extensions) and run with debugging.

A reference I used a lot in making this:
https://github.com/SpaceManiac/SpacemanDMM/wiki/Setting-up-Debugging

I've elected to add /world/Genesis() to enable debugging as soon as
possible, which allows debugging early init. This is taken directly from
tg.

A note on SpacemanDMM: I've only enabled the debugger here. We should
enable the (local on the users PC, NOT on github!) linter part of it
too, but that first requires dealing with a ton of linter issues,
largely from old tgcode.
Documentation for its config:
https://github.com/SpaceManiac/SpacemanDMM/blob/master/CONFIGURING.md

There is one thing that needs addressing either before or shortly after
this is merged- and that is explaining coders why their game suddenly
froze(The debugger defaults to breaking execution on runtimes).