[TM ONLY] Lowers pants 1 pixel down for males (#949)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Lowers pants for males of all races (ecxept elfs) 1 pixel down as it was
breaking some pants on them

Before - After

![image](https://github.com/user-attachments/assets/2ad024eb-d6ad-4de0-9144-4805dc5976af)

![image](https://github.com/user-attachments/assets/6d8d6b52-f3fe-416c-af7d-aca7acd29af1)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->