Bog ranger choice (#915)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
It makes the bog ranger a choice, instead of a randomly selected
It still has the same limit of only being >3 players
The first to choose the 3 rangers gets the role
Change the name of the bog crossbowman to ranger, since it fits and
females can get it

+ Removed females being forced into certain hairstyles, a bit weird to
have your hair changed as a vulp
+ Added a cloak to ranger for STYLE while they go alone in the
wilderness
+ Removed a bug where the crossbowman would get LEGENDARY SKILLS since
they got the skill points from both the bog guard and the crossbowman
skills
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Not having to play a ranger, if you do not want to for the shift
Having the choice to be something, so your character can be known as a
skilled bowman
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->