Chasuble and Stole are priestess compatible (#984)
## About The Pull Request

Priest additional cloak clothes are lacking any female sprites which is
making them appear as invisible onmob. This is PR is kinda fixes that by
giving this clothing pieces the same as male sprites. That serves as a
band-aid fix until someone fill look at chasuble as it looks 'meh' on
fem statics by being a little too chunky.


![image](https://github.com/user-attachments/assets/dd71e8f3-4ebe-4ff7-8757-841b548390e3)

![image](https://github.com/user-attachments/assets/68ebf166-683d-483d-98e5-417f5cfad03f)
In game screenshots for chasuble and stole on elf character.

Keep in mind that the same problem will _probably_ occur for dwarfs and
any onther small species as well, but they are out of the scope of this
PR as this is simple use of male sprites on females.

## Why It's Good For The Game

Invisible sprites are bad :(