Tiefling Customization - Horns + Wagging Tails (#978)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This gives Tieflings their own curated selection of many of the horn
options available. It also provides the framework for more custom horns
and tails in the future. Right now, only a pointed and a heart-shaped
tail tip are selectable, but it leaves room for future spriters!
![Screenshot 2024-09-12
181146](https://github.com/user-attachments/assets/44c65661-433d-4cda-8088-9ee8390618b2)
![Screenshot 2024-09-12
181227](https://github.com/user-attachments/assets/85e133dc-bd45-418a-857e-888761cb630a)
The tail wagging animation may be viewed below:
https://gyazo.com/0b8c6bfe8b3360020c679fab1a3b713d
https://gyazo.com/e23af536e53fcf7337a8260460589ef8
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Tieflings flicking their tails around is cute. More customization
options other than the hardcoded horns is nice for people to properly
express their characters.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>